Module zkevm_keccak::util::eth_types

Ethereum and Evm types used to deserialize responses from web3 / geth.

Structs

AccessList

Access list

Block

The block type returned from RPC calls. This is generic over a TX type which will be either the hash or the full transaction, i.e. Block<TxHash> or Block`.

Bytes

Wrapper type around Bytes to deserialize/serialize “0x” prefixed ethereum hex strings

H64

H160

Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

Signature

An ECDSA signature

Transaction

Details of a signed transaction

U64

Little-endian large integer type Unsigned 64-bit integer.

U256

Little-endian large integer type 256-bit unsigned integer.

Traits

Field

Trait used to reduce verbosity with the declaration of the [FieldExt] trait and its repr.

ToAddress

Trait used to convert a type to a Address.

ToBigEndian

Trait uset do convert a scalar value to a 32 byte array in big endian.

ToLittleEndian

Trait used to convert a scalar value to a 32 byte array in little endian.

ToScalar

Trait used to define types that can be converted to a 256 bit scalar value.

ToWord

Trait used to convert a type to a Word.

Type Definitions

Address

Hash

Ethereum Hash (256 bits).

Word

Ethereum Word (256 bits).