Module zkevm_keccak::util::expression

Modules

and

Returns 1 when expr[0] && expr[1] && ... == 1, and returns 0 otherwise. Inputs need to be boolean

not

Returns 1 when b == 0, and returns 0 otherwise. b needs to be boolean

or

Returns 1 when expr[0] || expr[1] || ... == 1, and returns 0 otherwise. Inputs need to be boolean

select

Returns when_true when selector == 1, and returns when_false when selector == 0. selector needs to be boolean.

sum

Returns the sum of the passed in cells

xor

Returns a ^ b. a and b needs to be boolean

Traits

Expr

Trait that implements functionality to get a constant expression from commonly used types.

Functions

expr_from_bytes

Given a bytes-representation of an expression, it computes and returns the single expression.

pow_of_two

Returns 2**by as FieldExt