Module zkevm_keccak::util::expression::select

Returns when_true when selector == 1, and returns when_false when selector == 0. selector needs to be boolean.

Functions

expr

Returns the when_true expression when the selector is true, else returns the when_false expression.

value

Returns the when_true value when the selector is true, else returns the when_false value.

value_word

Returns the when_true word when selector is true, else returns the when_false word.